[refactor] 주석 처리된 Order 디렉토리 삭제 및 security 디렉토리 common으로 이전
