--- src/main/java/com/example/baglemonster/order/dto/updateOrderRequestDto.java
@@ -1,9 +1,0 @@
-//package com.example.baglemonster.order.dto;
-//
-//import lombok.Getter;
-//
-//@Getter
-//public class updateOrderRequestDto {
-//    private Long productId;
-//    private Long quantity;
-//}
