--- src/main/java/com/example/baglemonster/order/entity/Order.java
@@ -1,50 +1,0 @@
-//package com.example.baglemonster.order.entity;
-//
-//import com.example.baglemonster.common.entity.Timestamped;
-//import com.example.baglemonster.store.entity.Store;
-//import com.example.baglemonster.user.entity.User;
-//import jakarta.persistence.*;
-//import lombok.*;
-//
-//@Builder
-//@Entity
-//@Getter
-//@NoArgsConstructor(access = AccessLevel.PROTECTED)
-//@AllArgsConstructor
-//public class Order extends Timestamped {
-//    /**
-//    * 컬럼 - 연관관계 컬럼을 제외한 컬럼을 정의합니다.
-//    */
-//    @Id
-//    @GeneratedValue(strategy = GenerationType.IDENTITY)
-//    private Long id;
-//
-//    private int totalPrice;
-//
-//    private String request;
-//
-//    private boolean status;
-//
-//    /**
-//    * 생성자 - 약속된 형태로만 생성가능하도록 합니다.
-//    */
-//
-//    /**
-//    * 연관관계 - Foreign Key 값을 따로 컬럼으로 정의하지 않고 연관 관계로 정의합니다.
-//    */
-//    @ManyToOne(fetch = FetchType.LAZY)
-//    @JoinColumn(name = "user_id", nullable = false)
-//    private User user;
-//
-//    @ManyToOne(fetch = FetchType.LAZY)
-//    @JoinColumn(name = "store_id", nullable = false)
-//    private Store store;
-//
-//    /**
-//    * 연관관계 편의 메소드 - 반대쪽에는 연관관계 편의 메소드가 없도록 주의합니다.
-//    */
-//
-//    /**
-//    * 서비스 메소드 - 외부에서 엔티티를 수정할 메소드를 정의합니다. (단일 책임을 가지도록 주의합니다.)
-//    */
-//}
